// frontend/src/environments/interface.ts
export interface Environment {
  production: boolean
  apiKey: string
  serverUrl: string
  authUrl: string
  registerUrl: string
  accountUrl: string
  updateUserUrl: string
  updatePasswordUrl: string
  personalBooksUrl: string
  booksUrl: string
  reviewsUrl: string
  resetPasswordUrl: string
  changePasswordUrl: string
  savePasswordUrl: string
}